JASPER-589
Implement POST and SUBSCRIBE processing logic and incorporate into Abe's changes

+++ jasper-platform/jsb/src/test/java/org/jasper/core/notification/TestTriggers.java
@@ -47,7 +47,7 @@ public class TestTriggers extends TestCase {
 		JsonResponseParser parser = new JsonResponseParser();
 		
 		// test passing in empty array
-		List<Integer> list = parser.parse(response, RURI);
+		List<Float> list = parser.parse(response, RURI);
 		
 		// test valid array and a bad room temperature (NaN)
 		response.add(tmp1);
